refactor: rename BatchOperationPageProcessor to BatchOperationChunkAppender
The class's primary responsibility is chunking items
 and appending them to the result buffer.
The new name better reflects what it does rather than
 what it operates on (pages).
